minimal modification in order to make keychain copying pipeline ready
