+++ src/main/java/com/sic/plugins/kpp/KPPKeychainsBuildWrapper.java
@@ -26,13 +26,11 @@ package com.sic.plugins.kpp;
 
 import com.sic.plugins.kpp.model.KPPKeychain;
 import com.sic.plugins.kpp.model.KPPKeychainCertificatePair;
+import hudson.EnvVars;
 import hudson.Extension;
 import hudson.FilePath;
 import hudson.Launcher;
-import hudson.model.AbstractBuild;
-import hudson.model.AbstractProject;
-import hudson.model.BuildListener;
-import hudson.model.Hudson;
+import hudson.model.*;
 import hudson.tasks.BuildWrapper;
 import hudson.tasks.BuildWrapperDescriptor;
 import java.io.File;
@@ -41,13 +39,15 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+
+import jenkins.tasks.SimpleBuildWrapper;
 import org.kohsuke.stapler.DataBoundConstructor;
 
 /**
  * Build wrapper for keychains
  * @author mb
  */
-public class KPPKeychainsBuildWrapper extends BuildWrapper {
+public class KPPKeychainsBuildWrapper extends SimpleBuildWrapper {
 
     private List<KPPKeychainCertificatePair> keychainCertificatePairs = new ArrayList<KPPKeychainCertificatePair>();
     private boolean deleteKeychainsAfterBuild;
@@ -92,20 +92,21 @@ public class KPPKeychainsBuildWrapper extends BuildWrapper {
     }
 
     @Override
-    public Environment setUp(AbstractBuild build, Launcher launcher, BuildListener listener) throws IOException, InterruptedException {
-        copyKeychainsToWorkspace(build);
-        return new EnvironmentImpl(keychainCertificatePairs);
+    public void setUp(Context context, Run<?, ?> run, FilePath filePath, Launcher launcher, TaskListener taskListener, EnvVars envVars) throws IOException, InterruptedException {
+        copyKeychainsToWorkspace(filePath);
+
+        Environment env = new EnvironmentImpl(keychainCertificatePairs);
+        env.buildEnvVars(context.getEnv());
+        context.setDisposer(new KPPKeychainsDisposer());
     }
 
     /**
      * Copy the keychains configured for this build job to the workspace of the job.
-     * @param build the current build
+     * @param projectWorkspace the current build
      * @throws IOException
      * @throws InterruptedException
      */
-    private void copyKeychainsToWorkspace(AbstractBuild build) throws IOException, InterruptedException {
-        FilePath projectWorkspace = build.getWorkspace();
-
+    private void copyKeychainsToWorkspace(FilePath projectWorkspace) throws IOException, InterruptedException {
         Hudson hudson = Hudson.getInstance();
         FilePath hudsonRoot = hudson.getRootPath();
 
@@ -148,7 +149,23 @@ public class KPPKeychainsBuildWrapper extends BuildWrapper {
     }
 
     /**
+     * Disposer class for cleaning up copied keychains
+     */
+    public class KPPKeychainsDisposer extends Disposer
+    {
+        @Override
+        public void tearDown(Run<?, ?> run, FilePath filePath, Launcher launcher, TaskListener taskListener) throws IOException, InterruptedException {
+            if (deleteKeychainsAfterBuild) {
+                for (FilePath keychainPath : copiedKeychains) {
+                    keychainPath.delete();
+                }
+            }
+        }
+    }
+
+    /**
      * Environment implementation that adds additional variables to the build.
+     * TODO: Does not need extend Environment anymore.
      */
     private class EnvironmentImpl extends Environment {
 
@@ -178,7 +195,6 @@ public class KPPKeychainsBuildWrapper extends BuildWrapper {
                     if (fileName!=null && fileName.length()!=0) {
                         String keychainPath = String.format("%s%s%s", env.get("WORKSPACE"), File.separator, fileName);
                         map.put(pair.getKeychainVariableName(), keychainPath);
-
                     }
                     if (password!=null && password.length()!=0)
                         map.put(pair.getKeychainPasswordVariableName(), keychain.getPassword());
@@ -193,17 +209,5 @@ public class KPPKeychainsBuildWrapper extends BuildWrapper {
         public void buildEnvVars(Map<String, String> env) {
             env.putAll(getEnvMap(env));
         }
-
-        @Override
-        public boolean tearDown(AbstractBuild build, BuildListener listener)
-                throws IOException, InterruptedException {
-            if (deleteKeychainsAfterBuild) {
-                for (FilePath filePath : copiedKeychains) {
-                    filePath.delete();
-                }
-            }
-            return true;
-        }
-
     }
 }
