+++ src/main/resources/com/sic/plugins/kpp/KPPKeychainsBuildWrapper/config.properties
@@ -1,6 +1,6 @@
 # The MIT License
 #
-# Copyright 2013 Michael B\u00e4r SIC! Software GmbH
+# Copyright 2013 michaelbar.
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -20,11 +20,11 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
-overwrite = Overwrite existing provisioning profiles
-delete = Delete copied provisioning profiles after build
-add_btn = Add Provisioning Profile
-profile = Provisioning Profile
-prefix = Variable Prefix
-variable = Variable
+overwrite = Overwrite existing keychains
+delete = Delete copied keychains after build
+add_btn = Add Keychain
+keychain = Keychain
+identity = Code Signing Identity
+prefix = Variables Prefix
+variables = Variables
 delete_btn = Delete
-
