+++ src/main/resources/com/sic/plugins/kpp/Messages.properties
@@ -20,14 +20,14 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
-KPPKeychainsBuildWrapper.DisplayName = Keychains and Code Signing Identities Pipeline
-KPPProvisioningProfilesBuildWrapper.DisplayName = Mobile Provisioning Profiles Pipeline
+KPPKeychainsBuildWrapper.DisplayName = Keychains and Code Signing Identities
+KPPProvisioningProfilesBuildWrapper.DisplayName = Mobile Provisioning Profiles
 KPPProvisioningProfilesBuildWrapper.NoProvisioningProfilesPathForMaster = The path to store mobile provisioning profile files on the master is not configured. Go the plugin main configuration page and give the path.
 KPPProvisioningProfilesBuildWrapper.NoProvisioningProfilesPathForSlave = The path to store mobile provisioning profile files on the slave is not configured. Go to the slave configuration page and give the path.
-KPPNodeProperty.DisplayName = Keychains and Provisioning Profiles Management Pipeline
+KPPNodeProperty.DisplayName = Keychains and Provisioning Profiles Management
 KPPManagementLink.Upload.Error.NoFile = Could not upload file! There was no file selected.
 KPPManagementLink.Upload.Error.WrongFileType = Wrong filetype! Uploaded file "%s" is no keychain or provisioning profile file.
-KPPManagementLink.DisplayName = Keychains and Provisioning Profiles Management Pipeline
-KPPKeychain.DisplayName = Keychain Pipeline
-KPPProvisioningProfile.DisplayName = Provisioning Profile Pipeline
+KPPManagementLink.DisplayName = Keychains and Provisioning Profiles Management
+KPPKeychain.DisplayName = Keychain
+KPPProvisioningProfile.DisplayName = Provisioning Profile
 KPPKeychainCertificatePair.DisplayName = ""
