--- src/main/resources/com/sic/plugins/kpp/simplebuildwrapper/KPPAddKeychains/config.jelly
@@ -1,58 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-The MIT License
-
-Copyright 2013 Michael Bär SIC! Software GmbH
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in
-all copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-THE SOFTWARE.
--->
-
-<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form">
-    <tr>
-        <td/>
-        <td colspan="2">
-            <div class="help" style="display:block;background-color:#F7F1DA;">
-                ${%HowToConfigureThisPlugin}
-            </div>
-        </td>
-        <td/>
-    </tr>
-        <f:entry field="keychain">
-            <label class="attach-previous">${%keychain}</label>
-            <f:textbox/>
-        </f:entry>
-        <f:entry field="codeSigningIdentity">
-            <label class="attach-previous">${%identity}</label>
-            <f:textbox/>
-        </f:entry>
-        <f:entry field="varPrefix">
-            <label class="attach-previous">${%prefix}</label>
-            <f:textbox/>
-        </f:entry>
-      <f:entry field="deleteKeychainsAfterBuild">
-        <f:checkbox/>
-        <label class="attach-previous">${%Delete Keychains after build?}</label>
-      </f:entry>
-      <f:entry field="overwriteExistingKeychains">
-          <f:checkbox/>
-            <label class="attach-previous">${%Overwrite existing keychains?}</label>
-      </f:entry>
-</j:jelly>
-
-
