--- src/main/resources/com/sic/plugins/kpp/simplebuildwrapper/KPPAddKeychains/config.properties
@@ -1,34 +1,0 @@
-# The MIT License
-#
-# Copyright 2013 michaelbar.
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in
-# all copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-# THE SOFTWARE.
-
-overwrite = Overwrite existing keychains
-delete = Delete copied keychains after build
-add_btn = Add Keychain
-keychain = Keychain
-identity = Code Signing Identity
-prefix = Variables Prefix
-variables = Variables
-delete_btn = Delete
-HowToConfigureThisPlugin= Build Wrapper for copying needed keychains for iOS Code Signing. Define keychain file name and \
-   code signing identitiy and make sure that those are already uploaded to Jenkins Server. The following variables will \
-  be added to your build environment: KEYCHAIN_PATH, KEYCHAIN_PASSWORD, CODE_SIGNING_IDENTITY
-
